Correction des operators

+++ Sources/Fonction_activations/Sigmoide.cpp
@@ -12,7 +12,7 @@ double Sigmoide::prim(double x) {
 }
 
 double Sigmoide::operator()(double x) {
-    double y = 1/(1+exp(-x));
+    double y = 1/(1+(double)exp(-x));
     return y;
 }
